# perf(sam): avoid searching for SAM CLI