--- a/src/shared/sam/cli/samCliSettings.ts
+++ b/src/shared/sam/cli/samCliSettings.ts
@@ -94,8 +94,10 @@ export class SamCliSettings extends fromExtensionManifest('aws.samcli', descript
     }
 
     /**
-     * Gets location of `sam` from user config, or tries to find `sam` on the
-     * system if the user config is invalid.
+     * Gets location of `sam` from:
+     * 1. previous saved location (if valid), or
+     * 2. user config (if valid), or
+     * 3. tries to find `sam` on the system if the user config is invalid.
      *
      * @returns `autoDetected=true` if auto-detection was _attempted_.
      */