--- a/src/shared/sam/cli/samCliSettings.ts
+++ b/src/shared/sam/cli/samCliSettings.ts
@@ -95,9 +95,9 @@ export class SamCliSettings extends fromExtensionManifest('aws.samcli', descript
 
     /**
      * Gets location of `sam` from:
-     * 1. previous saved location (if valid), or
-     * 2. user config (if valid), or
-     * 3. tries to find `sam` on the system if the user config is invalid.
+     * 1. user config (if any; overrides auto-detected location)
+     * 2. previous cached location (if valid)
+     * 3. searching for `sam` on the system
      *
      * @returns `autoDetected=true` if auto-detection was _attempted_.
      */