--- a/src/test/shared/sam/cli/samCliTestUtils.ts
+++ b/src/test/shared/sam/cli/samCliTestUtils.ts
@@ -21,9 +21,9 @@ export class MockSamCliProcessInvoker implements SamCliProcessInvoker {
 
         this.validateArgs(invokeSettings.arguments)
 
-        return ({
+        return {
             exitCode: 0,
-        } as any) as ChildProcessResult
+        } as any as ChildProcessResult
     }
 }