## Problem

- `SamCliLocationProvider.getLocation()` is called many times and always  searches for sam on the system, which is slow (especially on Windows).
- Hard to follow the logic of SamCliValidator, too much indirection  involving `isSamCliVersionCached()` and `getSamCliExecutableId()`.

## Solution
- Cache the last-found sam location in `DefaultSamCliLocationProvider`  instead of whatever `DefaultSamCliValidator` was trying to do.
- Use the cached location unless it fails `tryRun()`. This means Toolkit  still "shells out" to `sam`, which can be slow, but at least it avoids  a system search.

## Results

- Sample from CI shows that the first getLocation takes 800ms, later calls get a 5x speedup.
- Testing on Windows indicates first call can take 2-8(!) seconds, later calls are ~300ms.

```
2023-05-08 15:50:11 [VERBOSE]: samCliLocator: getLocation took 822ms
2023-05-08 15:50:16 [VERBOSE]: samCliLocator: getLocation took 119ms
2023-05-08 15:50:17 [VERBOSE]: samCliLocator: getLocation took 122ms
```

<!---
    REMINDER:
    - Read CONTRIBUTING.md first.
    - Add test coverage for your changes.
    - Update the changelog using `npm run newChange`.
    - Link to related issues/commits.
    - Testing: how did you test your changes?
    - Screenshots
-->

## License

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.